delete dir

+++ contracts/InternalContract.sol
@@ -12,32 +12,47 @@ contract SponsorWhitelistControl {
      * @dev get current Sponsored Balance for gas
      * @param contractAddr The address of the sponsored contract
      */
-    function getSponsoredBalanceForGas(address contractAddr) public returns (uint) {}
+    function getSponsoredBalanceForGas(address contractAddr)
+        public
+        returns (uint256)
+    {}
 
     /**
      * @dev get current Sponsored Gas fee upper bound
      * @param contractAddr The address of the sponsored contract
      */
-    function getSponsoredGasFeeUpperBound(address contractAddr) public returns (uint) {}
+    function getSponsoredGasFeeUpperBound(address contractAddr)
+        public
+        returns (uint256)
+    {}
 
     /**
      * @dev get collateral sponsor address
      * @param contractAddr The address of the sponsored contract
      */
-    function getSponsorForCollateral(address contractAddr) public returns (address) {}
+    function getSponsorForCollateral(address contractAddr)
+        public
+        returns (address)
+    {}
 
     /**
      * @dev get current Sponsored Balance for collateral
      * @param contractAddr The address of the sponsored contract
      */
-    function getSponsoredBalanceForCollateral(address contractAddr) public returns (uint) {}
+    function getSponsoredBalanceForCollateral(address contractAddr)
+        public
+        returns (uint256)
+    {}
 
     /**
      * @dev check if a user is in a contract's whitelist
      * @param contractAddr The address of the sponsored contract
      * @param user The address of contract user
      */
-    function isWhitelisted(address contractAddr, address user) public returns (bool) {}
+    function isWhitelisted(address contractAddr, address user)
+        public
+        returns (bool)
+    {}
 
     /**
      * @dev check if all users are in a contract's whitelist
@@ -51,20 +66,29 @@ contract SponsorWhitelistControl {
      * @param contractAddr The address of the sponsored contract
      * @param addresses The user address array
      */
-    function addPrivilegeByAdmin(address contractAddr, address[] memory addresses) public {}
+    function addPrivilegeByAdmin(
+        address contractAddr,
+        address[] memory addresses
+    ) public {}
 
     /**
      * @dev contract admin remove user from whitelist
      * @param contractAddr The address of the sponsored contract
      * @param addresses The user address array
      */
-    function removePrivilegeByAdmin(address contractAddr, address[] memory addresses) public {}
+    function removePrivilegeByAdmin(
+        address contractAddr,
+        address[] memory addresses
+    ) public {}
 
     // ------------------------------------------------------------------------
     // Someone will sponsor the gas cost for contract `contractAddr` with an
     // `upper_bound` for a single transaction.
     // ------------------------------------------------------------------------
-    function setSponsorForGas(address contractAddr, uint upper_bound) public payable {}
+    function setSponsorForGas(address contractAddr, uint256 upper_bound)
+        public
+        payable
+    {}
 
     // ------------------------------------------------------------------------
     // Someone will sponsor the storage collateral for contract `contractAddr`.
